backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
redis 
python-dotenv
aiofiles
pyjwt
numpy
requests
# For MySQL support (optional, defaults to SQLite if not configured)
pymysql
cryptography
